tasks/T01/api_client.py
#!/usr/bin/env python3
"""
API 客户端模块
根据老板提供的API文档实现具体接口
"""

import requests
import yaml
import logging
from typing import Dict, Any, Optional

logger = logging.getLogger(__name__)


class APIClient:
    """通用API客户端，具体实现待补充"""
    
    def __init__(self, config_path: str = "config.yaml"):
        with open(config_path, 'r', encoding='utf-8') as f:
            self.config = yaml.safe_load(f)
        
        self.endpoint = self.config['api']['endpoint']
        self.api_key = self.config['api']['api_key']
        self.timeout = self.config['api'].get('timeout', 30)
        
        self.session = requests.Session()
        if self.api_key:
            self.session.headers.update({'Authorization': f'Bearer {self.api_key}'})
    
    def get_market_data(self, symbols: list, start_date: str, end_date: str) -> Dict[str, Any]:
        """
        获取市场数据（日线、分钟线等）
        待根据API文档实现
        """
        raise NotImplementedError("等待老板提供API文档后实现")
    
    def get_stock_basic_info(self, symbols: list) -> Dict[str, Any]:
        """
        获取股票基本信息（市值、行业等）
        待实现
        """
        raise NotImplementedError("等待老板提供API文档后实现")
    
    def get_sector_data(self, sector_codes: list) -> Dict[str, Any]:
        """
        获取板块数据
        待实现
        """
        raise NotImplementedError("等待老板提供API文档后实现")
    
    def test_connection(self) -> bool:
        """测试API连接"""
        try:
            # 待实现具体测试接口
            return True
        except Exception as e:
            logger.error(f"API连接测试失败: {e}")
            return False


if __name__ == "__main__":
    # 简单测试
    client = APIClient()
    print("API客户端初始化完成（待实现具体接口）")