tasks/T01/requirements.txt
# T01 龙头战法选股依赖包

# 核心数据处理
pandas>=1.5.0
numpy>=1.21.0

# API请求
requests>=2.28.0

# 配置文件
pyyaml>=6.0

# tushare数据源 (必需)
tushare>=1.2.89

# 可选：技术指标计算
# ta-lib>=0.4.26  # 需要先安装TA-Lib C库

# 开发工具
black>=22.0.0
pylint>=2.15.0

# 时间处理
python-dateutil>=2.8.2